ai_microservices/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.23.2
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
xgboost==1.7.6
pydantic==2.9.2
joblib==1.3.2
pyarrow==15.0.2
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1
requests==2.32.3

